# Trabalho/swagger.yaml
openapi: 3.0.4
info:
  title: Green Herb - OpenAPI 3.0
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/
...
paths:
...
  /herb:
    post:
      tags:
        - herb
      summary: Create herb
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HerbCreate'
      responses:
        '200':
          description: Herb created
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
    get:
      tags:
        - herb
      summary: Get all herbs
      security:
        - accessToken: []
      responses:
        '200':
          description: List of herbs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Herb'
        '401':
          description: Unauthorized
  /herb/upload:
    post:
      tags:
        - herb
      summary: Upload herbs via CSV
      security:
        - accessToken: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Herbs created
        '400':
          description: Validation error (invalid CSV / duplicates / empty file)
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
        '500':
          description: Server error
  /herb/{id}:
    get:
      tags:
        - herb
      summary: Get herb by id
      security:
        - accessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Herb found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Herb'
        '400':
          description: Invalid id
        '401':
          description: Unauthorized
        '404':
          description: Herb not found
    patch:
      tags:
        - herb
      summary: Update herb by id
      security:
        - accessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HerbCreate'
      responses:
        '200':
          description: Herb updated successfully
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Herb not found
    delete:
      tags:
        - herb
      summary: Delete herb by id
      security:
        - accessToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Herb deleted
        '404':
          description: Herb not found
        '400':
          description: Herb is in use / invalid id
...
components:
  schemas:
...
    UserRef:
      type: object
      properties:
        id:
          type: string
          example: "6a03ebef990f3504734461ad"
        username:
          type: string
          format: string
          example: "ronaldo_cr7"
        email:
          type: string
          example: "ronaldo@alnassr.com"
      required: 
        - id
        - username
        - email
    HerbCreate:
      type: object
      properties: 
        name:
          type: string
          example: "Manjericão"
        description:
          type: string
          example: "Muito Utilizado em comidas"
        category:
          type: string
          enum: 
            - "aromática"
            - "culinaria"
            - "medicinal"
            - "outro"
          default: "aromática"
      required:
        - name
        - category
    Herb:
      type: object
      properties: 
        id: 
          type: string
          example: "6a06b273d7595b31b50e7450"
        name:
          type: string
          example: "Manjericão"
        description:
          type: string
          example: "Muito Utilizado em comidas"
        category:
          type: string
          enum: 
            - "aromática"
            - "culinaria"
            - "medicinal"
            - "outro"
          default: "aromática"
        createdBy:
          $ref: '#/components/schemas/UserRef'
        updatedBy:
          $ref: '#/components/schemas/UserRef'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - description
        - category
        - createdBy
...
    Error:
      type: object
      properties:
        type:
          type: string
        errors:
          type: object
          additionalProperties: 
            description: "Field: Error"
            type: string
      required:
        - type
        - errors
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT